Add Circle Icon shape overlay (RRO)
It is supposed to be the "device default" and fr/base is overlaid already
here in vendor/lineage as default, but I have experienced a situation, where
resetting to default (i.e. deactivate the respective icon shape RRO) has
left square shaped icons, so adding this as an extra option does not harm
and further allows a manual setting, if the "device default" looks weird.

Change-Id: I5e33b0315c2dee0495ee0194fe5206d23e3963a5